# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * QAID - AI-powered Financial Risk & Compliance API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface FileUploadInput {
  /** The ERP file to upload (Excel, CSV, or ZIP) */
  file: string;
}

export interface KpiSummary {
  overall_risk_score: number;
  risk_level: string;
  total_entries: number;
  flagged_accounts: number;
  suspicious_transactions: number;
  compliance_percentage: number;
  ai_confidence: number;
}

export interface RiskDistribution {
  level: string;
  count: number;
  percentage: number;
  color?: string;
}

export interface RiskTrend {
  period: string;
  avg_risk_score: number;
  entry_count: number;
  flagged_count?: number;
}

export interface RiskEntry {
  entry_id: string;
  account: string;
  /** @nullable */
  account_name?: string | null;
  /** @nullable */
  description?: string | null;
  amount: number;
  debit?: number;
  credit?: number;
  date: string;
  /** @nullable */
  user?: string | null;
  risk_score: number;
  fraud_probability: number;
  risk_level: string;
  triggered_rules: string[];
  ml_score?: number;
  rule_score?: number;
}

export interface RiskAccount {
  account_code: string;
  account_name?: string;
  total_entries: number;
  avg_risk_score: number;
  risk_level: string;
  total_amount: number;
  max_single_amount?: number;
}

export interface ChangedAccount {
  account_code: string;
  account_name: string;
  monthly_change_pct: number;
  current_month: number;
  previous_month: number;
}

export interface FraudIndicator {
  name: string;
  severity: string;
  count: number;
  description: string;
}

export interface ComplianceItem {
  standard: string;
  requirement: string;
  status: string;
  details: string;
  score?: number;
}

export interface ValidationIssue {
  type: string;
  severity: string;
  count: number;
  description: string;
}

export interface AnalysisResult {
  session_id: string;
  file_name: string;
  kpi_summary: KpiSummary;
  risk_distribution: RiskDistribution[];
  risk_trend: RiskTrend[];
  high_risk_entries: RiskEntry[];
  high_risk_accounts: RiskAccount[];
  top_changed_accounts: ChangedAccount[];
  fraud_indicators: FraudIndicator[];
  compliance_items: ComplianceItem[];
  recommendations: string[];
  ai_summary: string;
  validation_issues: ValidationIssue[];
  created_at: string;
  total_entries_analyzed?: number;
}

export interface TriggeredRule {
  rule_name: string;
  weight: number;
  description: string;
}

export type EntryDetailFeatures = {[key: string]: number};

export interface EntryDetail {
  entry_id: string;
  account: string;
  /** @nullable */
  account_name?: string | null;
  /** @nullable */
  description?: string | null;
  amount: number;
  debit?: number;
  credit?: number;
  date: string;
  /** @nullable */
  user?: string | null;
  risk_score: number;
  fraud_probability: number;
  risk_level: string;
  triggered_rules_detail: TriggeredRule[];
  ml_score: number;
  rule_score: number;
  recommendation: string;
  features?: EntryDetailFeatures;
}