# lib/api-zod/src/generated/types/changedAccount.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * QAID - AI-powered Financial Risk & Compliance API
 * OpenAPI spec version: 0.1.0
 */

export interface ChangedAccount {
  account_code: string;
  account_name: string;
  monthly_change_pct: number;
  current_month: number;
  previous_month: number;
}