# lib/api-zod/src/generated/types/entryDetail.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * QAID - AI-powered Financial Risk & Compliance API
 * OpenAPI spec version: 0.1.0
 */
import type { EntryDetailFeatures } from './entryDetailFeatures';
import type { TriggeredRule } from './triggeredRule';

export interface EntryDetail {
  entry_id: string;
  account: string;
  /** @nullable */
  account_name?: string | null;
  /** @nullable */
  description?: string | null;
  amount: number;
  debit?: number;
  credit?: number;
  date: string;
  /** @nullable */
  user?: string | null;
  risk_score: number;
  fraud_probability: number;
  risk_level: string;
  triggered_rules_detail: TriggeredRule[];
  ml_score: number;
  rule_score: number;
  recommendation: string;
  features?: EntryDetailFeatures;
}