# lib/api-zod/src/generated/types/fileUploadInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * QAID - AI-powered Financial Risk & Compliance API
 * OpenAPI spec version: 0.1.0
 */

export interface FileUploadInput {
  /** The ERP file to upload (Excel, CSV, or ZIP) */
  file: string;
}