# lib/api-zod/src/generated/types/kpiSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * QAID - AI-powered Financial Risk & Compliance API
 * OpenAPI spec version: 0.1.0
 */

export interface KpiSummary {
  overall_risk_score: number;
  risk_level: string;
  total_entries: number;
  flagged_accounts: number;
  suspicious_transactions: number;
  compliance_percentage: number;
  ai_confidence: number;
}