# lib/api-zod/src/generated/types/riskEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * QAID - AI-powered Financial Risk & Compliance API
 * OpenAPI spec version: 0.1.0
 */

export interface RiskEntry {
  entry_id: string;
  account: string;
  /** @nullable */
  account_name?: string | null;
  /** @nullable */
  description?: string | null;
  amount: number;
  debit?: number;
  credit?: number;
  date: string;
  /** @nullable */
  user?: string | null;
  risk_score: number;
  fraud_probability: number;
  risk_level: string;
  triggered_rules: string[];
  ml_score?: number;
  rule_score?: number;
}