# lib/api-zod/src/generated/types/riskTrend.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * QAID - AI-powered Financial Risk & Compliance API
 * OpenAPI spec version: 0.1.0
 */

export interface RiskTrend {
  period: string;
  avg_risk_score: number;
  entry_count: number;
  flagged_count?: number;
}