# lib/api-zod/src/generated/types/triggeredRule.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * QAID - AI-powered Financial Risk & Compliance API
 * OpenAPI spec version: 0.1.0
 */

export interface TriggeredRule {
  rule_name: string;
  weight: number;
  description: string;
}