# lib/api-zod/src/generated/types/validationIssue.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * QAID - AI-powered Financial Risk & Compliance API
 * OpenAPI spec version: 0.1.0
 */

export interface ValidationIssue {
  type: string;
  severity: string;
  count: number;
  description: string;
}